Provide example FASTA sequences in a compressed format
The large file size of our example sequence FASTA (>10 MB) prevents the ncov
repository from being used as a template repository on GitHub. This commit
replaces the uncompressed data with a gzip compressed file and updates the
corresponding tutorial documentation to include a decompression step before
users start running analyses.

An alternate solution to this one would be to not host the example data in the
repository at all but instead host it in an S3 bucket, etc. One benefit of the
current approach is that all of the data users need to run an analysis come with
the repository, making a complicated workflow just a little simpler.

Resolves #453